Added tests.
