+++ README.md
@@ -1,5 +1,8 @@
 # Sparta Social
 
+[![Test](https://github.com/dblock/sweatosphere/actions/workflows/test.yml/badge.svg)](https://github.com/dblock/sweatosphere/actions/workflows/test.yml)
+[![codecov](https://codecov.io/gh/dblock/sweatosphere/branch/main/graph/badge.svg)](https://codecov.io/gh/dblock/sweatosphere)
+
 Let's build a decentralized exercise activity tracker with [Atmosphere](https://atproto.com).
 
 ![](screenshot.png)
