+++ src/routes.ts
@@ -7,7 +7,7 @@ import { TID } from '@atproto/common'
 import { Agent, ComNS } from '@atproto/api'
 import express from 'express'
 import { getIronSession } from 'iron-session'
-import type { AppContext } from '#/index'
+import type { AppContext } from '#/server'
 import { home } from '#/pages/home'
 import { login } from '#/pages/login'
 import { env } from '#/lib/env'
