+++ tests/jest.env.ts
@@ -1,0 +1,5 @@
+process.env.NODE_ENV = 'test'
+process.env.PUBLIC_URL = ''
+process.env.DB_PATH = ":memory:"
+process.env.PORT = '50000'
+process.env.HOST = 'localhost'
